Implementacion_completa_apis_subida_archivos

+++ app/Http/Controllers/Aspirante/ExperienciaController.php
@@ -2,173 +2,213 @@
 
 namespace App\Http\Controllers\Aspirante;
 
-use App\Constants\ConstAgregarExperiencia\TiposExperiencia;
-use App\Constants\ConstAgregarExperiencia\TrabajoActual;
+
+use App\Http\Requests\RequestAspirante\RequestExperiencia\ActualizarExperienciaRequest;
 use Illuminate\Http\Request;
 use App\Models\Aspirante\Experiencia;
 use App\Models\Aspirante\Documento;// Importar el modelo Documento
 use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Facades\Log;
+
+use Illuminate\Support\Facades\DB;
+use App\Http\Requests\RequestAspirante\RequestExperiencia\CrearExperienciaRequest;
+
 class ExperienciaController
 {
     // Crear un registro de experiencia
-    public function crearExperiencia(Request $request)
+    public function crearExperiencia(CrearExperienciaRequest $request)
     {
-        // Validar los datos de entrada
-        $validator = Validator::make(request()->all(), [
-            'tipo_experiencia'             => 'required|string|in:'. implode(',', TiposExperiencia::all()),
-            'institucion_experiencia'      => 'required|string|min:3|max:100',
-            'cargo'                        => 'required|string|min:3|max:100',
-            'trabajo_actual'               => 'required|in:' . implode(',', TrabajoActual::all()),
-            'intensidad_horaria'           => 'nullable|integer|min:1|max:168',
-            'fecha_inicio'                 => 'nullable|date',// volver este campo a requerido
-            'fecha_finalizacion'           => 'nullable|date|after_or_equal:fecha_inicio',
-            'fecha_expedicion_certificado' => 'nullable|date',
-            'archivo'                      => 'required|file|mimes:pdf,jpg,png|max:2048', // Validación del archivo
-        ]);
+        try {
+            $experiencia = DB::transaction(function () use ($request) {
+                // Validar los datos de la experiencia
+                $datosExperiencia = $request->validated();
 
-        // Si la validación falla, se devuelve un mensaje de error
-        if ($validator->fails()) {
-            return response()->json($validator->errors()->toJson(), 400);
+                //crear experiencia
+                $experiencia = Experiencia::create($datosExperiencia);
+
+                // Verificar si se envió un archivo
+                if ($request->hasFile('archivo')) {
+                    $archivo = $request->file('archivo');
+                    $nombreArchivo = time() . '_' . $archivo->getClientOriginalName();
+                    $rutaArchivo = $archivo->storeAs('documentos/Experiencias', $nombreArchivo, 'public');
+
+                    // Guardar el documento relacionado con la experiencia
+                    Documento::create([
+                        'user_id'        => $request->user()->id, // Usuario autenticado
+                        'archivo'        => str_replace('public/', '', $rutaArchivo),
+                        'estado'         => 'pendiente',
+                        'documentable_id' => $experiencia->id_experiencia, // Relación polimórfica
+                        'documentable_type' => Experiencia::class,
+                    ]);
+                }
+
+                return $experiencia;
+            });
+
+            return response()->json([
+                'message' => 'Experiencia creada exitosamente',
+                'data'    => $experiencia
+            ], 201);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error al crear la experiencia o subir el archivo.',
+                'error'   => $e->getMessage()
+            ], 500);
         }
-
-        // Crear el registro de experiencia
-        $experiencia = Experiencia::create([
-            'tipo_experiencia'             => $request->input('tipo_experiencia'),
-            'institucion_experiencia'      => $request->input('institucion_experiencia'),
-            'cargo'                        => $request->input('cargo'),
-            'trabajo_actual'               => $request->input('trabajo_actual'),
-            'intensidad_horaria'           => $request->input('intensidad_horaria'),
-            'fecha_inicio'                 => $request->input('fecha_inicio'),
-            'fecha_finalizacion'           => $request->input('fecha_finalizacion'),
-            'fecha_expedicion_certificado' => $request->input('fecha_expedicion_certificado'),
-        ]);
-
-        // Verificar si se envió un archivo
-        if ($request->hasFile('archivo')) {
-            $archivo = $request->file('archivo');
-            $nombreArchivo = time() . '_' . $archivo->getClientOriginalName();
-            $rutaArchivo = $archivo->storeAs('documentos/Experiencias', $nombreArchivo, 'public');
-
-            // Guardar el documento relacionado con la experiencia
-            Documento::create([
-                'user_id'        => $request->user()->id, // Usuario autenticado
-                'archivo'        => str_replace('public/', '', $rutaArchivo),
-                'estado'         => 'pendiente',
-                'documentable_id' => $experiencia->id_experiencia, // Relación polimórfica
-                'documentable_type' => Experiencia::class,
-            ]);
-        }
-
-        // Devolver respuesta con el registro creado
-        return response()->json([
-            'message' => ' Experiencia y documento guardados correctamente',
-            'Experiencia'    => $experiencia
-        ], 201);
     }
     
     
     // Obtener todos los registros de experiencia
     public function obtenerExperiencias(Request $request)
     {
-        $user = $request->user(); // Obtiene el usuario autenticado
-    
-        // Verificar si el usuario está autenticado
-        if (!$user) {
-            return response()->json(['error' => 'Usuario no autenticado'], 401);
-        }
-    
-        // Obtener solo las experiencias que tienen documentos pertenecientes al usuario autenticado
-        $experiencias = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
-            $query->where('user_id', $user->id);
-        })->with(['documentosExperiencia' => function ($query) {
-            $query->select('id_documento', 'documentable_id', 'archivo', 'user_id'); // Relación polimórfica usa documentable_id
-        }])->get();
-    
-        // Agregar la URL del archivo a cada documento si existe
-        $experiencias->each(function ($experiencia) {
-            $experiencia->documentosExperiencia->each(function ($documento) {
-                if (!empty($documento->archivo)) {
-                    $documento->archivo_url = asset('storage/' . $documento->archivo);
-                }
+        try {
+            $user = $request->user(); // Obtiene el usuario autenticado
+
+            // Verificar si el usuario está autenticado
+            if (!$user) {
+                throw new \Exception('Usuario no autenticado', 401);
+            }
+
+            // Obtener solo las experiencias que tienen documentos pertenecientes al usuario autenticado
+            $experiencias = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
+                $query->where('user_id', $user->id);
+            })->with(['documentosExperiencia' => function ($query) {
+                $query->select('id_documento', 'documentable_id', 'archivo', 'user_id'); // Relación polimórfica usa documentable_id
+            }])
+            ->orderBy('created_at')
+            ->get();
+
+            // Verificar si se encontraron experiencias
+            if ($experiencias->isEmpty()) {
+                throw new \Exception ('No se encontraron experiencias', 404);
+            }
+
+            // Agregar la URL del archivo a cada documento si existe
+            $experiencias->each(function ($experiencia) {
+                $experiencia->documentosExperiencia->each(function ($documento) {
+                    if (!empty($documento->archivo)) {
+                        $documento->archivo_url = asset('storage/' . $documento->archivo);
+                    }
+                });
             });
-        });
+
+            return response()->json(['experiencias' => $experiencias], 200);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error al obtener las experiencias.',
+                'error'   => $e->getMessage()
+            ], $e->getCode() ?: 500);
+        }
+    }
+
     
-        return response()->json(['experiencias' => $experiencias], 200);
+    // Obtener un registro de experiencia por ID
+    public function obtenerExperienciasPorId(Request $request, $id)
+    {
+        try {
+            $user = $request->user(); // Obtiene el usuario autenticado
+
+            // Verificar si el usuario está autenticado
+            if (!$user) {
+                throw new \Exception('Usuario no autenticado', 401);
+            }
+
+            // Obtener solo las experiencias que tienen documentos pertenecientes al usuario autenticado
+            $experiencia = Experiencia::where('id_experiencia', $id) // Asegurar que use la clave primaria id_experiencia
+            ->whereHas('documentosExperiencia', function ($query) use ($user) {
+                $query->where('user_id', $user->id);
+                
+            })->with(['documentosExperiencia' => function ($query) {
+                $query->select('id_documento', 'documentable_id', 'archivo', 'user_id'); // Relación polimórfica usa documentable_id
+            }])
+            ->orderBy('created_at')
+            ->first();
+
+            // Verificar si se encontraron experiencias
+            if ($experiencia->isEmpty()) {
+                throw new \Exception ('No se encontraron experiencias', 404);
+            }
+
+            // Agregar la URL del archivo a cada documento si existe
+                $experiencia->documentosExperiencia->each(function ($documento) {
+                    if (!empty($documento->archivo)) {
+                        $documento->archivo_url = asset('storage/' . $documento->archivo);
+                    }
+                });
+
+            return response()->json(['experiencias' => $experiencia], 200);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error al obtener la experiencia.',
+                'error'   => $e->getMessage()
+            ], $e->getCode() ?: 500);
+        }
     }
     
 
     
     // Actualizar un registro de experiencia
-    public function actualizarExperiencia(Request $request, $id)
+    public function actualizarExperiencia(ActualizarExperienciaRequest $request, $id)
     {
-        $user = $request->user();
+        try {
+            $experiencia = DB::transaction(function () use ($request, $id) {
+                $user = $request->user();
 
-        // Buscar la experiencia que tenga documentos del usuario autenticado
-        $experiencia = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
-            $query->where('user_id', $user->id);
-        })->where('id_experiencia', $id)->firstOrFail(); // Asegurar que use la clave primaria id_experiencia
+                // Buscar la experiencia que tenga documentos del usuario autenticado
+                $experiencia = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
+                    $query->where('user_id', $user->id);
+                })->where('id_experiencia', $id)->firstOrFail(); // Asegurar que use la clave primaria id_experiencia
 
-        // Validar solo los campos que se envían en la solicitud
-        $validator = Validator::make($request->all(), [
-            'tipo_experiencia'             => 'sometimes|required|string|in:' . implode(',', TiposExperiencia::all()),
-            'institucion_experiencia'      => 'sometimes|required|string|min:3|max:100',
-            'cargo'                        => 'sometimes|required|string|min:3|max:100',
-            'trabajo_actual'               => 'sometimes|required|in:' . implode(',', TrabajoActual::all()),
-            'intensidad_horaria'           => 'sometimes|nullable|integer|min:1|max:168',
-            'fecha_inicio'                 => 'sometimes|required|date',
-            'fecha_finalizacion'           => 'sometimes|nullable|date|after_or_equal:fecha_inicio',
-            'fecha_expedicion_certificado' => 'sometimes|nullable|date',
-            'archivo'                      => 'sometimes|file|mimes:pdf,jpg,png|max:2048',
-        ]);
+                // Validar solo los campos que se envían en la solicitud
+                $datosExperienciaActualizar = $request->validated();
 
-        if ($validator->fails()) {
-            return response()->json($validator->errors()->toJson(), 400);
+                // Actualizar la experiencia
+                $experiencia->update($datosExperienciaActualizar);
+
+                // Manejo del archivo
+                if ($request->hasFile('archivo')) {
+                    $archivo = $request->file('archivo');
+                    $nombreArchivo = time() . '_' . $archivo->getClientOriginalName();
+                    $rutaArchivo = $archivo->storeAs('documentos/Experiencias', $nombreArchivo, 'public');
+
+                    // Buscar el documento asociado
+                    $documento = Documento::where('documentable_id', $experiencia->id_experiencia)
+                        ->where('documentable_type', Experiencia::class)
+                        ->where('user_id', $user->id)
+                        ->first();
+
+                    if ($documento) {
+                        Storage::disk('public')->delete($documento->archivo);
+                        $documento->update([
+                            'archivo' => str_replace('public/', '', $rutaArchivo),
+                            'estado'  => 'pendiente',
+                        ]);
+                    } else {
+                        Documento::create([
+                            'user_id'        => $user->id,
+                            'archivo'        => str_replace('public/', '', $rutaArchivo),
+                            'estado'         => 'pendiente',
+                            'documentable_id' => $experiencia->id_experiencia,
+                            'documentable_type' => Experiencia::class,
+                        ]);
+                    }
+                }
+                return $experiencia;
+            });
+
+            return response()->json([
+                'message' => 'Experiencia actualizada correctamente',
+                'data'    => $experiencia->fresh()
+            ], 200);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error al actualizar la experiencia o manejar el archivo.',
+                'error'   => $e->getMessage()
+            ], 500);
         }
-
-
-        // Actualizar los datos directamente
-        $data = $request->only([
-            'tipo_experiencia', 'institucion_experiencia', 'cargo', 'trabajo_actual',
-            'intensidad_horaria', 'fecha_inicio', 'fecha_finalizacion', 'fecha_expedicion_certificado'
-        ]);
-
-        $experiencia->update($data);
-
-        // Manejo del archivo
-        if ($request->hasFile('archivo')) {
-            $archivo = $request->file('archivo');
-            $nombreArchivo = time() . '_' . $archivo->getClientOriginalName();
-            $rutaArchivo = $archivo->storeAs('documentos/Experiencias', $nombreArchivo, 'public');
-
-            // Buscar el documento asociado
-            $documento = Documento::where('documentable_id', $experiencia->id_experiencia)
-                ->where('documentable_type', Experiencia::class)
-                ->where('user_id', $user->id)
-                ->first();
-
-            if ($documento) {
-                Storage::disk('public')->delete($documento->archivo);
-                $documento->update([
-                    'archivo' => str_replace('public/', '', $rutaArchivo),
-                    'estado'  => 'pendiente',
-                ]);
-            } else {
-                Documento::create([
-                    'user_id'        => $user->id,
-                    'archivo'        => str_replace('public/', '', $rutaArchivo),
-                    'estado'         => 'pendiente',
-                    'documentable_id' => $experiencia->id_experiencia,
-                    'documentable_type' => Experiencia::class,
-                ]);
-            }
-        }
-
-        return response()->json([
-            'message' => 'Experiencia actualizada correctamente',
-            'data'    => $experiencia->refresh()
-        ], 200);
     }
 
     
@@ -177,25 +217,40 @@ class ExperienciaController
     // Eliminar un registro de experiencia
     public function eliminarExperiencia(Request $request, $id)
     {
-        $user = $request->user();
+        try {
+            $user = $request->user();
 
-        // Buscar la experiencia que tenga documentos del usuario autenticado
-        $experiencia = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
-            $query->where('user_id', $user->id);
-        })->where('id_experiencia', $id)->firstOrFail(); // Asegurar que use la clave primaria id_experiencia
+            // Buscar la experiencia que tenga documentos del usuario autenticado
+            $experiencia = Experiencia::whereHas('documentosExperiencia', function ($query) use ($user) {
+                $query->where('user_id', $user->id);
+            })->where('id_experiencia', $id)->firstOrFail(); // Asegurar que use la clave primaria id_experiencia
 
-        // Eliminar el documento asociado
-        foreach ($experiencia->documentosExperiencia as $documento) {
-            // Eliminar el archivo del almacenamiento si existe
-            if (!empty($documento->archivo) && Storage::exists('public/' . $documento->archivo)) {
-                Storage::delete('public/' . $documento->archivo);
+            if(!$experiencia) {
+                return response()->json(['message' => 'Experiencia no encontrada'], 404);
             }
-            $documento->delete(); // Eliminar el documento de la base de datos
+
+            // Eliminar el documento asociado
+            DB::transaction(function () use ($experiencia){
+                foreach ($experiencia->documentosExperiencia as $documento) {
+                    // Eliminar el archivo del almacenamiento si existe
+                    if (!empty($documento->archivo) && Storage::exists('public/' . $documento->archivo)) {
+                        Storage::delete('public/' . $documento->archivo);
+                    }
+                    $documento->delete(); // Eliminar el documento de la base de datos
+                }
+                $experiencia->delete();
+            });
+
+            // Eliminar la experiencia
+            $experiencia->delete();
+
+            return response()->json(['message' => 'Experiencia eliminada correctamente'], 200);
+
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error al eliminar la experiencia.',
+                'error'   => $e->getMessage()
+            ], 500);
         }
-
-        // Eliminar la experiencia
-        $experiencia->delete();
-
-        return response()->json(['message' => 'Experiencia eliminada correctamente'], 200);
     }
 }
